Fixing bug in G6WriteIterator for N approaching 100,000

+++ c/graphLib/io/g6-read-iterator.h
@@ -29,8 +29,8 @@ extern "C"
 
         int order;
         int numCharsForOrder;
-        int numCharsForGraphEncoding;
-        int currGraphBuffSize;
+        size_t numCharsForGraphEncoding;
+        size_t currGraphBuffSize;
         char *currGraphBuff;
 
         graphP currGraph;
