+++ c/graphLib/io/g6-write-iterator.h
@@ -27,11 +27,11 @@ extern "C"
 
         int order;
         int numCharsForOrder;
-        int numCharsForGraphEncoding;
-        int currGraphBuffSize;
+        size_t numCharsForGraphEncoding;
+        size_t currGraphBuffSize;
         char *currGraphBuff;
 
-        int *columnOffsets;
+        size_t *columnOffsets;
 
         graphP currGraph;
     } G6WriteIterator;
